chore(tests): restructure integration tests by workflow
- Delete component-oriented `tests/integration/test_dataclass_loading.py`
- Move `DataClassConfigBuilder`-specific tests into workflow-based files:
  - 12 validation/coercion tests → `test_validation.py` (`TestDataClassValidation`)
  - 3 env coercion tests → `test_env_loading.py` (`TestDataClassENVCoercion`)
  - 1 multi-source priority test → `test_multi_source_merging.py`
- Replace duplicated `env` context managers in 4 files with shared `env_vars` fixture
- Add `SimpleConfigDC`, `DatabaseConfigDC`, `ComplexConfigDC` schemas to `conftest.py`
