+++ tests/integration/test_env_loading.py
@@ -6,35 +6,11 @@ handling, all value types, and the integer-before-float parse order fix
 from v0.1.1.
 """
 
-import os
-from contextlib import contextmanager
-
 import pytest
 
-from pyconfigre import ConfigBuilder
+from pyconfigre import ConfigBuilder, DataClassConfigBuilder
 
-from .conftest import AppConfig, SimpleConfig
-
-# ---------------------------------------------------------------------------
-# Helper
-# ---------------------------------------------------------------------------
-
-
-@contextmanager
-def _env(**kwargs):
-    """Set env vars for the duration of a with-block, then clean up."""
-    for k, v in kwargs.items():
-        os.environ[k] = v
-    try:
-        yield
-    finally:
-        for k in kwargs:
-            os.environ.pop(k, None)
-
-
-# ---------------------------------------------------------------------------
-# Tests
-# ---------------------------------------------------------------------------
+from .conftest import AppConfig, SimpleConfig, SimpleConfigDC, env_vars
 
 
 @pytest.mark.integration
@@ -43,7 +19,7 @@ class TestENVLoading:
 
     def test_flat_env_vars_with_prefix(self):
         """Prefixed env vars produce a flat config after prefix is stripped."""
-        with _env(
+        with env_vars(
             SVC_NAME="env-service",
             SVC_DEBUG="false",
             SVC_PORT="7070",
@@ -62,7 +38,7 @@ class TestENVLoading:
         This is the core fix for Issue-3 in v0.1.1. MYAPP__SERVER__PORT
         must produce config.server.port, not config["server__port"].
         """
-        with _env(
+        with env_vars(
             APP__APP_NAME="nested-service",
             APP__SERVER__HOST="10.0.0.1",
             APP__SERVER__PORT="9090",
@@ -90,7 +66,7 @@ class TestENVLoading:
         Uses a uniquely-prefixed canary key to avoid colliding with
         real process environment variables on any machine.
         """
-        with _env(PYCFG_CANARY_NAME="canary-value"):
+        with env_vars(PYCFG_CANARY_NAME="canary-value"):
             # Must be called inside the with block — the env var is only
             # set for the duration of the context manager.
             builder = ConfigBuilder(SimpleConfig).from_env()
@@ -101,7 +77,7 @@ class TestENVLoading:
 
     def test_strip_prefix_false_keeps_prefix_in_key(self):
         """strip_prefix=False retains the prefix as part of the key name."""
-        with _env(MYAPP_NAME="prefixed-app"):
+        with env_vars(MYAPP_NAME="prefixed-app"):
             builder = ConfigBuilder(SimpleConfig).from_env("MYAPP_", strip_prefix=False)
             raw = builder.peek()
 
@@ -110,7 +86,7 @@ class TestENVLoading:
 
     def test_lowercase_false_preserves_original_case(self):
         """lowercase=False keeps the original uppercase env var names."""
-        with _env(UPPER_NAME="case-test"):
+        with env_vars(UPPER_NAME="case-test"):
             builder = ConfigBuilder(SimpleConfig).from_env("UPPER_", lowercase=False)
             raw = builder.peek()
 
@@ -119,7 +95,7 @@ class TestENVLoading:
 
     def test_nested_false_keeps_double_underscore_flat(self):
         """nested=False prevents __ from being treated as a separator."""
-        with _env(FLAT__DB__HOST="some-host"):
+        with env_vars(FLAT__DB__HOST="some-host"):
             builder = ConfigBuilder(SimpleConfig).from_env("FLAT__", nested=False)
             raw = builder.peek()
 
@@ -129,7 +105,7 @@ class TestENVLoading:
 
     def test_all_value_types_parsed(self):
         """Env var string values are coerced to their correct Python types."""
-        with _env(
+        with env_vars(
             T_NAME="type-test",
             T_DEBUG="true",
             T_PORT="42",
@@ -151,7 +127,7 @@ class TestENVLoading:
         This is the parse-order fix in v0.1.1: int() is attempted before
         float() so that "8000" → 8000, not 8000.0.
         """
-        with _env(INT_PORT="8000"):
+        with env_vars(INT_PORT="8000"):
             builder = ConfigBuilder(SimpleConfig).from_env("INT_")
             raw = builder.peek()
 
@@ -160,3 +136,81 @@ class TestENVLoading:
             f"Expected int but got {type(raw['port']).__name__}. "
             "Check that int() is attempted before float() in _parse_value."
         )
+
+
+@pytest.mark.integration
+class TestDataClassENVCoercion:
+    """DataClassConfigBuilder string coercion from environment variables."""
+
+    def test_env_strings_coerced_to_field_types(self):
+        """Environment variables are strings; dataclass builder coerces them."""
+        with env_vars(
+            DCSVC_NAME="env-service",
+            DCSVC_DEBUG="true",
+            DCSVC_PORT="7070",
+            DCSVC_API_KEY="env-key-abc",
+        ):
+            config = DataClassConfigBuilder(SimpleConfigDC).from_env("DCSVC_").build()
+
+        assert config.name == "env-service"
+        assert config.debug is True
+        assert isinstance(config.debug, bool)
+        assert config.port == 7070
+        assert isinstance(config.port, int)
+        assert config.api_key == "env-key-abc"
+
+    def test_falsy_bool_strings_from_env(self):
+        """Falsy string values are coerced to False for bool fields."""
+        with env_vars(
+            DCSVC_NAME="env-service",
+            DCSVC_DEBUG="no",
+            DCSVC_PORT="7070",
+        ):
+            config = DataClassConfigBuilder(SimpleConfigDC).from_env("DCSVC_").build()
+
+        assert config.debug is False
+
+    def test_nested_env_vars_instantiate_dataclasses(self):
+        """Double-underscore env vars produce nested dataclass instances."""
+        from dataclasses import dataclass, field
+
+        @dataclass
+        class ServerConfigDC:
+            host: str = "0.0.0.0"
+            port: int = 8000
+            debug: bool = False
+
+        @dataclass
+        class DatabaseConfigDC:
+            name: str
+            username: str
+            password: str
+            host: str = "localhost"
+            port: int = 5432
+
+        @dataclass
+        class AppConfigDC:
+            database: DatabaseConfigDC
+            app_name: str = "myapp"
+            version: str = "0.1.0"
+            server: ServerConfigDC = field(default_factory=ServerConfigDC)
+
+        with env_vars(
+            DCAPP__APP_NAME="nested-service",
+            DCAPP__SERVER__HOST="10.0.0.1",
+            DCAPP__SERVER__PORT="9090",
+            DCAPP__SERVER__DEBUG="false",
+            DCAPP__DATABASE__HOST="db.internal",
+            DCAPP__DATABASE__PORT="5432",
+            DCAPP__DATABASE__NAME="prod",
+            DCAPP__DATABASE__USERNAME="svc",
+            DCAPP__DATABASE__PASSWORD="secret",
+        ):
+            config = DataClassConfigBuilder(AppConfigDC).from_env("DCAPP__").build()
+
+        assert config.app_name == "nested-service"
+        assert config.server.host == "10.0.0.1"
+        assert config.server.port == 9090
+        assert config.server.debug is False
+        assert config.database.host == "db.internal"
+        assert config.database.name == "prod"
